==> pruebas/1.ProcesadoDatasets.ipynb <==
# Procesado de los datasets inciales 
Primer vistazo a los datasets iniciales y preprocesado inicial para trabajar con ellos posteriormente. 

---

import numpy as np
import pandas as pd

## 1. Importar datos
- **clean_users**: genero, edad, categoria, regularidad, id
- **Prado_artworks_wikidata**: cuadros
- **Prado_emotions**: user, cuadro, emotion, polaridad
- **Prado_users**: id, edad, genero, pais

clean_users_data = "data/clean_users.csv"
cuadros_data = "data/Prado_artworks_wikidata.csv"
emotions_data = "data/Prado_emotions.csv"
prado_users_data = "data/Prado_users.csv"

clean_users  = pd.read_csv(clean_users_data)
cuadros = pd.read_csv(cuadros_data)
emotions = pd.read_csv(emotions_data)
prado_users = pd.read_csv(prado_users_data)

print(clean_users.shape)
clean_users.head(3)

print(cuadros.shape)
cuadros.head(3)

print(emotions.shape)
emotions.head(3)

print(prado_users.shape)
prado_users.head(3)

## 2. Procesar datos

#### - Reducir paises a regiones:
> Ya que el dataset de prado_users tiene pais vamos a utilizar el dataset "countries of the world" para identificar las regiones de las que proceden

countries_data = "data/countries_of_the_world.csv"
countries = pd.read_csv(countries_data)
countries = countries.filter(items = ['Country', 'Region'])

countries.head()

# # Eliminar el espacio en blanco despues de cada nombre de pais
# for i in range(0, countries.shape[0] - 1):
#     countries.loc[i] = countries.loc[i].replace(countries.loc[i]['Country'], countries.loc[i]['Country'][:-1])

# countries.to_csv("data/countries_of_the_world.csv")

# ## Sustituiremos cada pais de procedencia de prado_users por un numero q corresponda con su region
# # Ver cuantas regiones hay
# num_regions = 0
# for i in countries['Region'].unique():
#     num_regions += 1
#     countries = countries.replace(i,num_regions)

# Formatear los nombres de USA y Argentina
prado_users = prado_users.replace('USA','United States')
prado_users = prado_users.replace('Argenitina','Argentina')

# Reemplazar los paises por el numero de regiones
for country in prado_users['country'].unique():
    if country == 'Other':
        prado_users = prado_users.replace(country,0) #sustituir por 0 los usuarios que no sabemos de donde son
    else:
        prado_users = prado_users.replace(country,countries.loc[countries['Country'] == country]['Region'].to_list()[0])

countries.loc[countries['Country'] == 'United States']['Region'].to_list()[0]

prado_users.head()

#### - Formatear rangos de edades:
> Asignar a cada rango de edad un número, igual que hicimos en el paso anterior con las regiones

# Formatear los <12 y >70
prado_users = prado_users.replace('<12', '0-12')
prado_users = prado_users.replace('>70', '70-100')

# Calcularel numero correspondiente a cada rango de edad
age_index = 1

for age_range in prado_users.sort_values(by=['age'])['age'].unique():
    prado_users = prado_users.replace(age_range, age_index)
    age_index += 1
age_index -=1

prado_users.head()

#### - Formatear género:
> Cambiar el género por un número

prado_users = prado_users.replace('female', 0)
prado_users = prado_users.replace('male', 1)

prado_users.head()

## 3. Normalizar datos


from sklearn.preprocessing import StandardScaler

users_scaled = prado_users.copy()
users_scaled.drop('userId', axis=1, inplace=True) # No queremos normalizar el id (no es necesario para el clustering)

scaler = StandardScaler(with_mean=False, with_std=True)

users_scaled = pd.DataFrame(data=scaler.fit_transform(users_scaled), columns=users_scaled.columns.to_list())

users_scaled

users_scaled.insert(0, 'userId', prado_users.userId)

users_scaled

users_scaled.to_csv("data/Prado_users_scaled.csv", index= 0)
