==> pruebas/2.UserEmotionsList.ipynb <==
# Lista de emociones de cada usuario
Añadir al dataset de usuarios que cuadros les ha generado a cada usuario emociones positivas, negativas y mixtas

---

import pandas as pd
import numpy as np

prado_emotions_data = 'data/Prado_emotions.csv'
prado_emotions = pd.read_csv(prado_emotions_data)
prado_emotions.head(5)

# Para evitar que haya mas de una emocion por cuadro por usuario calculamos la media de las polaridades 
# que tiene cada usuario por cada cuadro
p = prado_emotions.copy()
# p['Polarity'] = np.where(p['Polarity'] == 'positive', 1, p['Polarity'])
# p['Polarity'] = np.where(p['Polarity'] == 'negative', 0, p['Polarity'])
# p['Polarity'] = np.where(p['Polarity'] == 'mixed', 0.5, p['Polarity'])
p['Polarity'] = np.select([(p['Polarity'] == 'positive'),(p['Polarity'] == 'negative'),(p['Polarity'] == 'mixed')] ,[1,0,0.5])

p.groupby(['userId','artworkId'])['Polarity'].mean()

emotions=pd.DataFrame(p.groupby(['userId','artworkId'])['Polarity'].mean().reset_index())
# No tenemos en cuenta mixed
# emotions['Polarity'] = np.select([(emotions['Polarity'] >= .5),(emotions['Polarity'] < .5)],  ['positive', 'negative'])
# tenemos en cuenta mixed
emotions['Polarity'] = np.select([(emotions['Polarity'] > .5),(emotions['Polarity'] < .5), (emotions['Polarity']==.5)],  ['positive', 'negative','mixed'])

emotions.head(5)

### Crear dataset con usuarios y listas de sus cuadros positivos, negativos y mixed 
users_emotions_list = {}
i = 1

for id in list(prado_emotions.userId.unique()):
    user = prado_emotions[prado_emotions['userId'] == id]
    
    pos_artworks = list(user[user['Polarity'] == 'positive'].artworkId)
    neg_artworks = list(user[user['Polarity'] == 'negative'].artworkId)
    mix_artworks = list(user[user['Polarity'] == 'mixed'].artworkId)
    
    users_emotions_list.update({i: [id, pos_artworks, neg_artworks, mix_artworks]})
    i += 1
    
users_emotions_df = pd.DataFrame.from_dict(users_emotions_list, orient='index', columns=['userId', 'positive', 'negative', 'mixed'])
users_emotions_df = users_emotions_df.sort_values(by=['userId'])
users_emotions_df = users_emotions_df.set_index(pd.Index([i for i in range(0,len(users_emotions_df))]))

# Crear dataset de usuarios y listas de sus cuadros positivos(solo una polaridad por cuadro)
users_emotions_list = {}
i = 1

for id in list(emotions.userId.unique()):
    user = emotions[emotions['userId'] == id]
    
    pos_artworks = list(user[user['Polarity'] == 'positive'].artworkId)
    neg_artworks = list(user[user['Polarity'] == 'negative'].artworkId)
    mix_artworks = list(user[user['Polarity'] == 'mixed'].artworkId)
    
    users_emotions_list.update({i: [id, pos_artworks, neg_artworks, mix_artworks]})
    i += 1
    
users_emotions_df = pd.DataFrame.from_dict(users_emotions_list, orient='index', columns=['userId', 'positive', 'negative', 'mixed'])
users_emotions_df = users_emotions_df.sort_values(by=['userId'])
users_emotions_df = users_emotions_df.set_index(pd.Index([i for i in range(0,len(users_emotions_df))]))

users_emotions_df

prado_users = pd.read_csv('data/Prado_users.csv')

pos = users_emotions_df['positive']
neg = users_emotions_df['negative']
mix = users_emotions_df['mixed']

prado_users = prado_users.join(pos)
prado_users = prado_users.join(neg)
prado_users = prado_users.join(mix)

prado_users.to_csv('data/Prado_users_emotions.csv', index=0)

prado_users = pd.read_csv('data/Prado_users.csv')

pos = users_emotions_df['positive']
neg = users_emotions_df['negative']
mix = users_emotions_df['mixed']

prado_users = prado_users.join(pos)
prado_users = prado_users.join(neg)
prado_users = prado_users.join(mix)

prado_users.to_csv('data/Prado_users_emotions_OnePolarity.csv', index=0)

prado_users_scaled = pd.read_csv('data/Prado_users_scaled.csv')

pos = users_emotions_df['positive']
neg = users_emotions_df['negative']
mix = users_emotions_df['mixed']

prado_users_scaled = prado_users_scaled.join(pos)
prado_users_scaled = prado_users_scaled.join(neg)
prado_users_scaled = prado_users_scaled.join(mix)

prado_users_scaled.to_csv('data/Prado_users_emotions_scaled.csv', index=0)

prado_users_scaled = pd.read_csv('data/Prado_users_scaled.csv')

pos = users_emotions_df['positive']
neg = users_emotions_df['negative']
mix = users_emotions_df['mixed']

prado_users_scaled = prado_users_scaled.join(pos)
prado_users_scaled = prado_users_scaled.join(neg)
prado_users_scaled = prado_users_scaled.join(mix)

prado_users_scaled.to_csv('data/Prado_users_emotions_scaled_OnePolarity.csv', index=0)

prado_users.loc(0)[0]

prado_users = pd.read_csv('data/Prado_users_emotions.csv')
prado_users['positive'] = prado_users['positive'].apply(eval)
for art1 in prado_users.loc(0)[0].positive:
    print(art1)
