==> pruebas/3.CalculoSimilitudes.ipynb <==
# Calculo de las matrices de similitud
Calcular las similitudes entre obras y entre usuarios y generar matrices de similitud para usarlas posteriormente

---

import numpy as np
import pandas as pd
import src.SimilarityUsers as su
import src.SimilarityArtworks as sa

## 1. Similitud Cuadros

artworks_csv = pd.read_csv(sa.ARTWORKS_CSV)
artworks_matrix = np.zeros((len(artworks_csv),len(artworks_csv)))

i = 0
for a in artworks_csv['ID']:
    j = 0
    for b in artworks_csv['ID']:
        artworks_matrix[i][j] = sa.SimilarityArtworks().computeSimilarity(a, b)
        j +=1
    i+=1

artworks_sim = pd.DataFrame(artworks_matrix, index = [i for i in artworks_csv['ID']], columns = [i for i in artworks_csv['ID']])
artworks_sim.to_csv('data/Prado_artworks_similarity.csv')

## 2. Similitud usuarios

artworks_sim = pd.read_csv('data/Prado_artworks_similarity.csv', index_col = 0)

users_csv = pd.read_csv(su.USERS_EMOTIONS_CSV)
# Tratar las listas de emociones para que no se lean como strings
users_csv['positive'] = users_csv['positive'].apply(eval)
users_csv['negative'] = users_csv['negative'].apply(eval)
users_csv['mixed'] = users_csv['mixed'].apply(eval)
users_matrix = []

SimUsers = su.SimilarityUsers(age_index = np.count_nonzero(users_csv['age'].unique()), artworks_sim = artworks_sim)

users_csv

for i in range(0, len(users_csv)):
  sim_list = []
  for j in range(0, len(users_csv)):
    sim = SimUsers.computeSimilarity(users_csv.loc(0)[i]['userId'], users_csv.loc(0)[j]['userId'])
    if sim >= 0:
      sim_list.append(sim)
    else:
      sim_list.append(0)
      
  users_matrix.append(sim_list)

users_sim = pd.DataFrame(users_matrix, index = [i for i in users_csv['userId']], columns = [i for i in users_csv['userId']])
users_sim.to_csv('data/Prado_users_similarity_OnePolarity.csv')

### Anexo: Similitud de individuos explicadores
Similitud con los individuos explicadores de cada cluster, generados externamenteprado_users = pd.read_csv('data/Prado_users_emotions.csv')prado_users = pd.read_csv('data/Prado_users_emotions.csv')

artworks_sim = pd.read_csv('data/Prado_artworks_similarity.csv', index_col = 0)

users_csv = pd.read_csv('data/Prado_users_individuo_explicador.csv')
# Tratar las listas de emociones para que no se lean como strings
users_csv['positive'] = users_csv['positive'].apply(eval)
users_csv['negative'] = users_csv['negative'].apply(eval)
users_csv['mixed'] = users_csv['mixed'].apply(eval)
users_matrix = []

SimUsers = su.SimilarityUsers(age_index = np.count_nonzero(users_csv['age'].unique()), artworks_sim = artworks_sim)

users_csv

for i in range(0, len(users_csv)):
  sim_list = []
  for j in range(0, len(users_csv)):
    sim = SimUsers.computeSimilarity(users_csv.loc(0)[i]['userId'], users_csv.loc(0)[j]['userId'])
    if sim >= 0:
      sim_list.append(sim)
    else:
      sim_list.append(0)
      
  users_matrix.append(sim_list)

users_sim = pd.DataFrame(users_matrix, index = [i for i in users_csv['userId']], columns = [i for i in users_csv['userId']])
users_sim.to_csv('data/Prado_users_explicador_similarity.csv')
