==> pruebas/5.IndividuoExplicador.ipynb <==
# Estudio de la clasficacion en clusters de los usuarios

import pandas as pd
import numpy as np

users = pd.read_csv('data/Prado_users_clusters_OnePolarity.csv')
emotions = pd.read_csv('data/Prado_users_emotions_OnePolarity.csv')
artworks = pd.read_csv('data/Prado_artworks_wikidata.csv')

users['positive'] = emotions['positive'].apply(eval)
users['negative'] = emotions['negative'].apply(eval)
users['mixed'] = emotions['mixed'].apply(eval)

users

clusters = []
for c in users.cluster.unique():
  clusters.append(users[users['cluster'] == c])

atr = clusters[0].columns[1]
clusters[0][atr].mode()[0]

clusters[1].describe()

import collections

users_promedio = pd.read_csv('data/Prado_users_emotions_scaled.csv')

for i, c in enumerate(clusters):
    print("-----------------------------------------------------------------------------------------------------")
    print('Cluster {}:'.format(c.cluster.iloc[0]))
    
    g_arr = []
    for g in c.gender.unique():
        g_arr.append(np.count_nonzero(c['gender'] == g))
        # g = 'male' if count_nonzero(c['gender'] == 'male') > np.count_nonzero(c['gender'] == 'female') else 'female'
    gender_max = c.gender.unique()[g_arr.index(max(g_arr))]
    
    age_arr = []
    for a  in c.age.unique():
        age_arr.append(np.count_nonzero(c['age'] == a))
    age_max = c.age.unique()[age_arr.index(max(age_arr))]
    
    country_arr = []
    for co in c.country.unique():
        country_arr.append(np.count_nonzero(c['country'] == co))
    country_max = c.country.unique()[country_arr.index(max(country_arr))]
    
    positive_arr = []
    for l in c.positive:
      for k in l:
        # title = artworks[artworks['ID'] == k]['Title'].iloc[0]
        # positive_arr.append(title)
        positive_arr.append(k)
        
    negative_arr = []
    for l in c.negative:
      for k in l:
        # title = artworks[artworks['ID'] == k]['Title'].iloc[0]
        # negative_arr.append(title)
        negative_arr.append(k)
        
    mixed_arr = []
    for l in c.mixed:
      for k in l:
        # title = artworks[artworks['ID'] == k]['Title'].iloc[0]
        # mixed_arr.append(title)
        mixed_arr.append(k)
    
    #######
    uProm = {"userId":'expl'+str(i),"age":2.933903194819822,"gender":0.0 if gender_max == 'female' else 2.021722956442927,"country":1.7874048008113603,"positive":positive_arr[:5],"negative":negative_arr[:5],"mixed":mixed_arr[:5]}
    users_promedio = users_promedio.append(uProm,ignore_index=True)
    #######
    
    print("numero de gente en cluster", i, sum(g_arr))
    print("Genero con mas aparicion en cluster", i,gender_max, "con un ", max(g_arr)/sum(g_arr)*100,"%")
    print("Edad con mas aparicion en cluster ", i, age_max , "con un ", max(age_arr)/sum(age_arr)*100,"%")
    print("Pais con mas aparicion en cluster ", i, country_max, "con un ", max(country_arr)/sum(country_arr)*100,"%")
    
    #-------------------------------------------------
    
    # sorted(zip(positive_arr, [c for c in collections.Counter(positive_arr).values()]), key=lambda x : x[1], reverse=True)
    pos = sorted(dict(collections.Counter(positive_arr)).items(), key=lambda item: item[1], reverse=True)
    
    pos_titles = [artworks[artworks['ID'] == id].Title.item() for id in [p[0] for p in pos]]
    print("\nCuadros más gustados: ", list(zip(pos_titles, pos)))
    
    pos_artists = [artworks[artworks['ID'] == id].Artist.item() for id in [p[0] for p in pos]]
    sorted_pos_artists = sorted(dict(collections.Counter(pos_artists)).items(), key=lambda item: item[1], reverse=True)
    print("\nArtistas más gustados: ", sorted_pos_artists)
    
    pos_categories = [artworks[artworks['ID'] == id].Category.item() for id in [p[0] for p in pos]]
    sorted_pos_categories = sorted(dict(collections.Counter(pos_categories)).items(), key=lambda item: item[1], reverse=True)
    print("\nCategorías más gustadas: ", sorted_pos_categories)
   
    #--------------------------------------------------
    
    neg = sorted(dict(collections.Counter(negative_arr)).items(), key=lambda item: item[1], reverse=True)
    
    neg_titles = [artworks[artworks['ID'] == id].Title.item() for id in [n[0] for n in neg]]
    print("\nCuadros menos gustados: ", list(zip(neg_titles, neg)))
    
    neg_artists = [artworks[artworks['ID'] == id].Artist.item() for id in [n[0] for n in neg]]
    sorted_neg_artists = sorted(dict(collections.Counter(neg_artists)).items(), key=lambda item: item[1], reverse=True)
    print("\nArtistas menos gustados: ", sorted_neg_artists)
    
    neg_categories = [artworks[artworks['ID'] == id].Category.item() for id in [n[0] for n in neg]]
    sorted_neg_categories = sorted(dict(collections.Counter(neg_categories)).items(), key=lambda item: item[1], reverse=True)
    print("\nCategorías menos gustadas: ", sorted_neg_categories)
    
    #---------------------------------------------------
    
#     mix = sorted(dict(collections.Counter(mixed_arr)).items(), key=lambda item: item[1], reverse=True)
    
#     mix_titles = [artworks[artworks['ID'] == id].Title.item() for id in [m[0] for m in mix]]
#     print("\nCuadros mixed: ", list(zip(mix_titles, mix)))
    
#     mix_artists = [artworks[artworks['ID'] == id].Artist.item() for id in [m[0] for m in mix]]
#     sorted_mix_artists = sorted(dict(collections.Counter(mix_artists)).items(), key=lambda item: item[1], reverse=True)
#     print("\nArtistas mixed: ", sorted_mix_artists)
    
#     mix_categories = [artworks[artworks['ID'] == id].Category.item() for id in [m[0] for m in mix]]
#     sorted_mix_categories = sorted(dict(collections.Counter(mix_categories)).items(), key=lambda item: item[1], reverse=True)
#     print("\nCategorías mixed: ", sorted_mix_categories)
  
  
    #---------------------------------------------------
    
    print("----------------------------------------------------------------------------------------------------\n")

users_promedio.to_csv('data/Prado_users_individuo_explicador.csv', index=0)

users_promedio

users[users['medoid'] != -1]

users_promedio.append(users[users['medoid'] != -1]).sort_values(by=['cluster'])
